Fix problems with spec runner page and Grid
specBundle is not being built at all and specs.html is set to actually render
the app. Set the main elements to dynamically add in both the spec page and the main page
and set invisible in the case of the spec page. Also, add a dependency for a Underscore and
create functions in Util, existy and truthy, based on Fogus in Functional Javascript. Continue
work in the grid on multi-square not selection.